python/01-tutorials/05-agentkit-memory/mem0_memory/pyproject.toml
[project]
name = "mem0-simple"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk==1.21.0",
    "veadk-python[extensions]==0.5.20",
    "veadk-python==0.5.20",
    "agentkit-sdk-python==0.4.3",
    "mem0ai==0.1.118",
]
